Moved QueryResult to Cypher
