+++ community/bolt/src/test/java/org/neo4j/bolt/v1/messaging/util/MessageMatchers.java
@@ -49,6 +49,8 @@ import org.neo4j.bolt.v1.messaging.message.SuccessMessage;
 import org.neo4j.bolt.v1.packstream.BufferedChannelInput;
 import org.neo4j.bolt.v1.packstream.BufferedChannelOutput;
 import org.neo4j.bolt.v1.transport.integration.TestNotification;
+import org.neo4j.cypher.internal.javacompat.BaseToObjectValueWriter;
+import org.neo4j.cypher.result.QueryResult;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Notification;
 import org.neo4j.graphdb.Relationship;
@@ -56,8 +58,6 @@ import org.neo4j.graphdb.spatial.Point;
 import org.neo4j.kernel.api.exceptions.Status;
 import org.neo4j.kernel.impl.util.HexPrinter;
 import org.neo4j.values.AnyValue;
-import org.neo4j.values.BaseToObjectValueWriter;
-import org.neo4j.values.result.QueryResult;
 import org.neo4j.values.virtual.MapValue;
 
 import static org.hamcrest.CoreMatchers.containsString;
