+++ community/bolt/src/test/java/org/neo4j/bolt/v1/runtime/spi/StreamMatchers.java
@@ -27,8 +27,8 @@ import java.util.Arrays;
 import java.util.Iterator;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import org.neo4j.cypher.result.QueryResult;
 import org.neo4j.values.AnyValue;
-import org.neo4j.values.result.QueryResult;
 import org.neo4j.values.storable.NumberValue;
 
 import static java.util.Arrays.asList;
