+++ community/cypher/cypher/src/main/java/org/neo4j/cypher/internal/javacompat/ExecutionResult.java
@@ -29,6 +29,7 @@ import java.util.stream.Stream;
 
 import org.neo4j.cypher.CypherException;
 import org.neo4j.cypher.internal.InternalExecutionResult;
+import org.neo4j.cypher.result.QueryResult;
 import org.neo4j.graphdb.ExecutionPlanDescription;
 import org.neo4j.graphdb.Notification;
 import org.neo4j.graphdb.QueryExecutionException;
@@ -40,7 +41,6 @@ import org.neo4j.graphdb.ResourceIterator;
 import org.neo4j.graphdb.Result;
 import org.neo4j.kernel.impl.query.QueryExecutionEngine;
 import org.neo4j.kernel.impl.query.QueryExecutionKernelException;
-import org.neo4j.values.result.QueryResult;
 
 /**
  * Holds Cypher query result sets, in tabular form. Each row of the result is a map
