+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/v3_2/ExecutionResultWrapper.scala
@@ -39,13 +39,13 @@ import org.neo4j.cypher.internal.frontend.v3_2.SemanticDirection.{BOTH, INCOMING
 import org.neo4j.cypher.internal.frontend.v3_2.notification.{DeprecatedPlannerNotification, InternalNotification, PlannerUnsupportedNotification, RuntimeUnsupportedNotification, _}
 import org.neo4j.cypher.internal.frontend.v3_3
 import org.neo4j.cypher.internal.javacompat.ValueUtils
+import org.neo4j.cypher.result.QueryResult
+import org.neo4j.cypher.result.QueryResult.Record
 import org.neo4j.graphdb
 import org.neo4j.graphdb.Result.{ResultRow, ResultVisitor}
 import org.neo4j.graphdb.impl.notification.{NotificationCode, NotificationDetail}
 import org.neo4j.graphdb.{Notification, ResourceIterator}
 import org.neo4j.values.AnyValue
-import org.neo4j.values.result.QueryResult
-import org.neo4j.values.result.QueryResult.Record
 
 import scala.collection.JavaConverters._
