+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/v3_3/runtime/helpers/RuntimeJavaValueConverter.scala
@@ -22,8 +22,8 @@ package org.neo4j.cypher.internal.compatibility.v3_3.runtime.helpers
 import java.util.{List => JavaList, Map => JavaMap}
 
 import org.neo4j.cypher.internal.frontend.v3_3.helpers.Eagerly.immutableMapValues
+import org.neo4j.cypher.result.QueryResult.{QueryResultVisitor, Record}
 import org.neo4j.values.AnyValue
-import org.neo4j.values.result.QueryResult.{QueryResultVisitor, Record}
 
 import scala.collection.JavaConverters._
 import scala.collection.Map
