+++ enterprise/cypher/cypher/src/test/scala/org/neo4j/cypher/internal/compiled_runtime/v3_3/codegen/CompiledExecutionResultTest.scala
@@ -34,11 +34,11 @@ import org.neo4j.cypher.internal.frontend.v3_3.test_helpers.CypherFunSuite
 import org.neo4j.cypher.internal.javacompat.{ResultRecord, ValueUtils}
 import org.neo4j.cypher.internal.spi.v3_3.QueryContext
 import org.neo4j.cypher.internal.v3_3.executionplan.GeneratedQueryExecution
+import org.neo4j.cypher.result.QueryResult.QueryResultVisitor
 import org.neo4j.graphdb.NotFoundException
 import org.neo4j.graphdb.Result.{ResultRow, ResultVisitor}
 import org.neo4j.helpers.collection.Iterators
 import org.neo4j.values.AnyValue
-import org.neo4j.values.result.QueryResult.QueryResultVisitor
 import org.neo4j.values.storable._
 import org.neo4j.values.virtual.{ListValue, MapValue}
